### 2017-11-30 #478 @whedon
[PRE REVIEW]: Gophernotes: the Go kernel for Jupyter and nteract
**Submitting author:** @dwhitena (<a href="http://orcid.org/0000-0003-1155-8889">Daniel Whitenack</a>)
**Repository:** <a href="https://github.com/gopherdata/gophernotes" target ="_blank">https://github.com/gopherdata/gophernotes</a>
**Version:** v1.0.0-RC
