### 2017-12-08 #478 @Kevin-Mattheus-Moerman
Hey Lucas @lgeiger, is this your cup of tea? Are you able to review this submission for JOSS?
