### 2017-12-08 #478 @Kevin-Mattheus-Moerman
@dwhitena our system has not detected your paper.md and paper.bib files. Can you please add these to your repository [as per our guidelines](http://joss.theoj.org/about#author_guidelines). Thanks
