### 2017-12-10 #478 @arfon
@arokem - would you be willing to edit this submission for JOSS?
