### 2017-12-11 #478 @dwhitena
@arokem I would suggest Sebastien Binet (`sbinet`) and/or Brendan Tracey (`btracey`). I'm traveling again today, but hoping to resolve the issue with the included paper today or tomorrow morning at the latest. Thanks again!
