### 2017-12-13 #478 @dwhitena
@arokem The issues with the paper and metadata files should be resolved. Let me know if they still aren't detected. Thanks!
