### 2017-12-15 #508 @whedon
[REVIEW]: Gophernotes: the Go kernel for Jupyter and nteract
**Submitting author:** @dwhitena (<a href="http://orcid.org/0000-0003-1155-8889">Daniel Whitenack</a>)
**Repository:** <a href="https://github.com/gopherdata/gophernotes" target ="_blank">https://github.com/gopherdata/gophernotes</a>
**Version:** v1.0.0-RC

- [x] **Version:** Does the release version given match the GitHub release (v1.0.0-RC)?
