### 2017-12-15 #508 @arfon
@dwhitena - this should fix the errors compiling your paper: https://github.com/gopherdata/gophernotes/pull/92
