### 2018-01-24 #508 @vsivsi
I’ve worked my way through some of the checklist.  I actually had an [open issue](https://github.com/gopherdata/gophernotes/issues/17#issuecomment-355446855) on the project repo from over a year ago, and I did an ‘@‘ comment to the author asking about progress on that and haven’t received any response since I sent it several weeks ago.  I’ll ping him again and see if I have any better luck.  @arokem
