### 2018-02-12 #508 @vsivsi
@arokem Working on it now.  Sorry for delay, it slipped into a busy period the past few weeks, but this week is pretty clear, so it will be done by Friday.
