### 2018-02-16 gopherdata/gophernotes#101 @vsivsi
Include available DOIs for references in JOSS submission
Hi @dwhitena

In finishing up my [review of your JOSS submission](https://github.com/openjournals/joss-reviews/issues/508), it appears that at least [one of your cited references has a DOI available](http://ebooks.iospress.nl/publication/42900) that is not included in the `.bib` file in this repo.

The [JOSS reviewer checklist](https://github.com/openjournals/joss-reviews/issues/508#issue-282288672) asks:

> Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
