+++ main.go
@@ -11,7 +11,7 @@ const (
 	Version string = "1.0.0"
 
 	// ProtocolVersion defines the Jupyter protocol version.
-	ProtocolVersion string = "5.0"
+	ProtocolVersion string = "5.1"
 )
 
 func main() {
