### 2018-04-07 404fc47
simplify doEval()
