+++ kernel.go
@@ -424,6 +424,10 @@ func doEval(ir *interp.Interp, code string) (val []interface{}, err error) {
 	// Check if the last node is an expression. If the last node is not an expression then nothing
 	// is returned as a value. For example evaluating a function declaration shouldn't return a value but
 	// just have the side effect of declaring the function.
+	//
+	// This is actually needed only for gomacro classic interpreter
+	// (the fast interpreter already returns values only for expressions)
+	// but retained for compatibility.
 	var srcEndsWithExpr bool
 	if len(nodes) > 0 {
 		_, srcEndsWithExpr = nodes[len(nodes)-1].(ast.Expr)
@@ -433,18 +437,11 @@ func doEval(ir *interp.Interp, code string) (val []interface{}, err error) {
 	compiledSrc := ir.CompileAst(srcAst)
 
 	// Evaluate the code.
-	result, results := ir.RunExpr(compiledSrc)
+	results := base.PackValues(ir.RunExpr(compiledSrc))
 
 	// If the source ends with an expression, then the result of the execution is the value of the expression. In the
 	// event that all return values are nil, the result is also nil.
 	if srcEndsWithExpr {
-		// `len(results) == 0` implies a single result stored in `result`.
-		if len(results) == 0 {
-			if val := base.ValueInterface(result); val != nil {
-				return []interface{}{val}, nil
-			}
-			return nil, nil
-		}
 
 		// Count the number of non-nil values in the output. If they are all nil then the output is skipped.
 		nonNilCount := 0
@@ -460,7 +457,6 @@ func doEval(ir *interp.Interp, code string) (val []interface{}, err error) {
 		if nonNilCount > 0 {
 			return values, nil
 		}
-		return nil, nil
 	}
 
 	return nil, nil
