### 2018-04-07 3ddbe1f
doEval(): preallocate a slice of values with required length
