### 2018-04-07 e7a9663
display images: show as image each expression result that implements Go interface 'image.Image'
