+++ README.md
@@ -188,16 +188,10 @@ $ docker run -it -p 8888:8888 -v /path/to/local/notebooks:/path/to/notebooks/in/
 gophernotes uses [gomacro](https://github.com/cosmos72/gomacro) under the hood to evaluate Go code interactively. You can evaluate most any Go code with gomacro, but there are some limitation, which are discussed in further detail [here](https://github.com/cosmos72/gomacro#current-status).  Most noteably, gophernotes does NOT support:
 
 - third party packages when running natively on Mac and Windows - This is a current limitation of the Go `plugin` package.
-- unexported struct fields
-- interfaces - They can be declared, but nothing more: there is no way to implement them or call their methods
-- extracting methods from types - For example time.Duration.String should return a func(time.Duration) string but currently gives an error. Instead extracting methods from objects is supported: time.Duration(1s).String correctly returns a func() string
-- goto
-- named return values
-- named imports like:
-
-    ```
-    import tf "github.com/tensorflow/tensorflow/tensorflow/go"
-    ```
+- unexported struct fields are emulated
+- interpreted interfaces - They can be declared, but nothing more: there is no way to implement them or call their methods
+  Note: interfaces imported from standard library or third party packages are compiled, and fully functional.
+- goto is only partially implemented
 
 ## Troubleshooting
 
