### 2018-05-28 f696d76
Inject a display function into the user space as well as the runtime interfaces
