+++ runtime/x_package.go
@@ -1,0 +1,157 @@
+// this file was generated by gomacro command: import _i "github.com/gopherdata/gophernotes/runtime"
+// DO NOT EDIT! Any change will be lost when the file is re-generated
+
+package runtime
+
+import (
+	r "reflect"
+
+	"github.com/cosmos72/gomacro/imports"
+)
+
+// reflection: allow interpreted code to import "github.com/gopherdata/gophernotes/runtime"
+func init() {
+	imports.Packages["github.com/gopherdata/gophernotes/runtime"] = imports.Package{
+		Binds: map[string]r.Value{
+			"HTML":               r.ValueOf(HTML),
+			"JPEG":               r.ValueOf(JPEG),
+			"JSON":               r.ValueOf(JSON),
+			"JavaScript":         r.ValueOf(JavaScript),
+			"Latex":              r.ValueOf(Latex),
+			"MIMETypeHTML":       r.ValueOf(MIMETypeHTML),
+			"MIMETypeJPEG":       r.ValueOf(MIMETypeJPEG),
+			"MIMETypeJSON":       r.ValueOf(MIMETypeJSON),
+			"MIMETypeJavaScript": r.ValueOf(MIMETypeJavaScript),
+			"MIMETypeLatex":      r.ValueOf(MIMETypeLatex),
+			"MIMETypeMarkdown":   r.ValueOf(MIMETypeMarkdown),
+			"MIMETypePNG":        r.ValueOf(MIMETypePNG),
+			"MIMETypeSVG":        r.ValueOf(MIMETypeSVG),
+			"Markdown":           r.ValueOf(Markdown),
+			"Math":               r.ValueOf(Math),
+			"PNG":                r.ValueOf(PNG),
+			"Render":             r.ValueOf(Render),
+			"SVG":                r.ValueOf(SVG),
+			"Text":               r.ValueOf(Text),
+		}, Types: map[string]r.Type{
+			"DisplayData":          r.TypeOf((*DisplayData)(nil)).Elem(),
+			"HTMLRenderable":       r.TypeOf((*HTMLRenderable)(nil)).Elem(),
+			"JPEGRenderable":       r.TypeOf((*JPEGRenderable)(nil)).Elem(),
+			"JSONRenderable":       r.TypeOf((*JSONRenderable)(nil)).Elem(),
+			"JavaScriptRenderable": r.TypeOf((*JavaScriptRenderable)(nil)).Elem(),
+			"LatexRenderable":      r.TypeOf((*LatexRenderable)(nil)).Elem(),
+			"MIMEBundleRenderable": r.TypeOf((*MIMEBundleRenderable)(nil)).Elem(),
+			"MarkdownRenderable":   r.TypeOf((*MarkdownRenderable)(nil)).Elem(),
+			"PNGRenderable":        r.TypeOf((*PNGRenderable)(nil)).Elem(),
+			"SVGRenderable":        r.TypeOf((*SVGRenderable)(nil)).Elem(),
+		}, Proxies: map[string]r.Type{
+			"HTMLRenderable":       r.TypeOf((*P_github_com_gopherdata_gophernotes_runtime_HTMLRenderable)(nil)).Elem(),
+			"JPEGRenderable":       r.TypeOf((*P_github_com_gopherdata_gophernotes_runtime_JPEGRenderable)(nil)).Elem(),
+			"JSONRenderable":       r.TypeOf((*P_github_com_gopherdata_gophernotes_runtime_JSONRenderable)(nil)).Elem(),
+			"JavaScriptRenderable": r.TypeOf((*P_github_com_gopherdata_gophernotes_runtime_JavaScriptRenderable)(nil)).Elem(),
[…]
