### 2018-05-28 eaa6814
Evaluate one statement at a time to better handle method declarations in gomacro
