+++ main.go
@@ -6,12 +6,11 @@ import (
 )
 
 const (
-
 	// Version defines the gophernotes version.
-	Version string = "1.0.0"
+	Version = "1.0.0"
 
 	// ProtocolVersion defines the Jupyter protocol version.
-	ProtocolVersion string = "5.1"
+	ProtocolVersion = "5.1"
 )
 
 func main() {
