### 2018-05-28 c2c5ae6
Include empty metadata bundle instead of null in display data
