### 2018-05-28 b5d88f9
Render the execution result like calling Display if there is a single result.
