### 2018-05-30 710233f
Lock writes to the sockets to fix a bug with interleaved publishes.
