+++ kernel.go
@@ -39,14 +39,20 @@ type ConnectionInfo struct {
 	IP              string `json:"ip"`
 }
 
+// Socket wraps a zmq socket with a lock which should be used to control write access.
+type Socket struct {
+	Socket *zmq.Socket
+	Lock   *sync.Mutex
+}
+
 // SocketGroup holds the sockets needed to communicate with the kernel,
 // and the key for message signing.
 type SocketGroup struct {
-	ShellSocket   *zmq.Socket
-	ControlSocket *zmq.Socket
-	StdinSocket   *zmq.Socket
-	IOPubSocket   *zmq.Socket
-	HBSocket      *zmq.Socket
+	ShellSocket   Socket
+	ControlSocket Socket
+	StdinSocket   Socket
+	IOPubSocket   Socket
+	HBSocket      Socket
 	Key           []byte
 }
 
@@ -88,6 +94,13 @@ const (
 	kernelIdle     = "idle"
 )
 
+// RunWithSocket invokes the `run` function after acquiring the `Socket.Lock` and releases the lock when done.
+func (s *Socket) RunWithSocket(run func(socket *zmq.Socket) error) error {
+	s.Lock.Lock()
+	defer s.Lock.Unlock()
+	return run(s.Socket)
+}
+
 // runKernel is the main entry point to start the kernel.
 func runKernel(connectionFile string) {
 
@@ -124,9 +137,9 @@ func runKernel(connectionFile string) {
 	// TODO gracefully shutdown the heartbeat handler on kernel shutdown by closing the chan returned by startHeartbeat.
 
 	poller := zmq.NewPoller()
-	poller.Add(sockets.ShellSocket, zmq.POLLIN)
-	poller.Add(sockets.StdinSocket, zmq.POLLIN)
-	poller.Add(sockets.ControlSocket, zmq.POLLIN)
+	poller.Add(sockets.ShellSocket.Socket, zmq.POLLIN)
+	poller.Add(sockets.StdinSocket.Socket, zmq.POLLIN)
+	poller.Add(sockets.ControlSocket.Socket, zmq.POLLIN)
 
 	// msgParts will store a received multipart message.
 	var msgParts [][]byte
@@ -144,8 +157,8 @@ func runKernel(connectionFile string) {
 			switch socket := item.Socket; socket {
 
 			// Handle shell messages.
-			case sockets.ShellSocket:
-				msgParts, err = sockets.ShellSocket.RecvMessageBytes(0)
+			case sockets.ShellSocket.Socket:
+				msgParts, err = sockets.ShellSocket.Socket.RecvMessageBytes(0)
 				if err != nil {
 					log.Println(err)
 				}
@@ -159,12 +172,12 @@ func runKernel(connectionFile string) {
 				handleShellMsg(ir, msgReceipt{msg, ids, sockets})
 
 				// TODO Handle stdin socket.
-			case sockets.StdinSocket:
-				sockets.StdinSocket.RecvMessageBytes(0)
+			case sockets.StdinSocket.Socket:
+				sockets.StdinSocket.Socket.RecvMessageBytes(0)
 
 				// Handle control messages.
-			case sockets.ControlSocket:
-				msgParts, err = sockets.ControlSocket.RecvMessageBytes(0)
+			case sockets.ControlSocket.Socket:
+				msgParts, err = sockets.ControlSocket.Socket.RecvMessageBytes(0)
 				if err != nil {
 					log.Println(err)
 					return
@@ -197,46 +210,51 @@ func prepareSockets(connInfo ConnectionInfo) (SocketGroup, error) {
 
 	// Create the shell socket, a request-reply socket that may receive messages from multiple frontend for
[…]
