### 2018-06-02 720be35
update vendored copy of gomacro; update kernel.go for API change in gomacro/fast/Interp.RunExpr()
