### 2018-06-02 c75e6bc
simplify @SpencerPark branch https://github.com/SpencerPark/gophernotes/tree/mime-rendering to implement rendering of HTML, JSON, JPEG, PNG, SVG, LaTeX... New functions: display.PNG([]byte), display.JPEG([]byte), display.HTML(string), display.SVG(string) ...
