### 2018-06-02 be43b70
merge @SpencerPark commit https://github.com/gopherdata/gophernotes/commit/710233fed632bf3f38d7dc1d25cb2db59f6d7ccc
From SpencerPark <spinnr95@gmail.com>:
    Lock writes to the sockets to fix a bug with interleaved publishes.
