### 2018-06-04 bfa25f8
As per @SpencerPark recommendation, split the 'func Display(data DisplayData) error' function that sends multimedia to Jupyter for displaying from the various factory functions 'func display.*(...) DisplayData' that create and return DisplayData values. Also remove the global variable globalReceipt and use a closure instead.
