+++ display.go
@@ -3,7 +3,6 @@ package main
 import (
 	"errors"
 	"fmt"
-	"image"
 	r "reflect"
 	"strings"
 
@@ -14,102 +13,139 @@ import (
 // See http://ipython.readthedocs.io/en/stable/api/generated/IPython.display.html#IPython.display.display
 // for a good overview of the support types. Note: This is missing _repr_markdown_ and _repr_javascript_.
 
-var globalReceipt *msgReceipt // ugly global variable. any alternative?
-
 const (
-	mimeTypeHTML       = "text/html"
-	mimeTypeMarkdown   = "text/markdown"
-	mimeTypeLatex      = "text/latex"
-	mimeTypeSVG        = "image/svg+xml"
-	mimeTypePNG        = "image/png"
-	mimeTypeJPEG       = "image/jpeg"
-	mimeTypeJSON       = "application/json"
-	mimeTypeJavaScript = "application/javascript"
+	MIMETypeHTML       = "text/html"
+	MIMETypeJavaScript = "application/javascript"
+	MIMETypeJPEG       = "image/jpeg"
+	MIMETypeJSON       = "application/json"
+	MIMETypeLatex      = "text/latex"
+	MIMETypeMarkdown   = "text/markdown"
+	MIMETypePNG        = "image/png"
+	MIMETypePDF        = "application/pdf"
+	MIMETypeSVG        = "image/svg+xml"
 )
+
+// injected as placeholder in the interpreter, it's then replaced at runtime
+// by a closure that knows how to talk with Jupyter
+func stubDisplay(DisplayData) error {
+	return errors.New("cannot display: connection with Jupiter not registered")
+}
 
 // TODO handle the metadata
 
-func render2(mimeType string, data interface{}) error {
-	return render3(mimeType, fmt.Sprint(data), data)
-}
-
-func render3(mimeType string, text string, data interface{}) error {
-	receipt := globalReceipt
-	if receipt == nil {
-		return errors.New("msgReceipt is nil, cannot send display_data message")
-	}
-	return receipt.PublishDisplayData(
-		bundledMIMEData{
-			"text/plain": text,
+func MakeDisplayData(mimeType string, data interface{}) DisplayData {
+	return DisplayData{
+		Data: BundledMIMEData{
+			"text/plain": fmt.Sprint(data),
 			mimeType:     data,
-		}, make(bundledMIMEData))
+		},
+	}
 }
 
-func HTML(html string) error {
-	return render2(mimeTypeHTML, html)
+func MakeDisplayData3(mimeType string, plaintext string, data interface{}) DisplayData {
+	return DisplayData{
+		Data: BundledMIMEData{
+			"text/plain": plaintext,
+			mimeType:     data,
+		},
+	}
 }
 
-func Markdown(markdown string) error {
-	return render2(mimeTypeMarkdown, markdown)
+func Bytes(mimeType string, bytes []byte) DisplayData {
+	return MakeDisplayData3(mimeType, mimeType, bytes)
 }
 
-func SVG(svg string) error {
-	return render2(mimeTypeSVG, svg)
+func HTML(html string) DisplayData {
+	return MakeDisplayData(MIMETypeHTML, html)
 }
 
-func PNG(png []byte) error {
-	return render3(mimeTypePNG, "{png-image}", png) // []byte are encoded as base64 by the marshaller
+func JSON(json map[string]interface{}) DisplayData {
+	return MakeDisplayData(MIMETypeJSON, json)
 }
 
-func JPEG(jpeg []byte) error {
-	return render3(mimeTypeJPEG, "{jpeg-image}", jpeg) // []byte are encoded as base64 by the marshaller
[…]
