### 2018-06-09 a0b6c36
update Limitations in README.md
