### 2018-06-10 568d6d7
automatically display image.Image or display.Data returned by REPL only if it's the only non-nil result
