### 2018-06-10 bd0dab5
take updated README.md from gopherdata/gophernotes master branch
