### 2018-06-10 9e18fed
if the only non-nil result returned by REPL is an image.Image or display.Data, publish it with "execute_result" Jupyter message, not with "display_data"
