### 2018-06-14 43d3181
add file examples/Display.ipynb and link it in README.md
