### 2018-06-18 123f3ee
document Mac OS X requirement of Go 1.10.2+ to import third party libraries at runtime
