+++ README.md
@@ -65,7 +65,8 @@ $ jupyter --data-dir
 
 ### Mac
 
-**Important Note** - gomacro relies on the `plugin` package when importing third party libraries.  This package is only supported on Linux currently.  Thus, if you need to utilize third party packages in your Go notebooks and you are running on Mac, you should use the [Docker](#docker) install and run gophernotes/Jupyter in Docker.
+**Important Note** - gomacro relies on the `plugin` package when importing third party libraries. This package works reliably on Mac OS X only with Go 1.10.2+ as long as you **never** execute the command `strip gophernotes`.
+If you can only compile gophernotes with Go <= 1.10.1 on Mac, consider using the [Docker](#docker) install and run gophernotes/Jupyter in Docker.
 
 ```sh
 $ go get github.com/gopherdata/gophernotes
@@ -88,7 +89,7 @@ $ jupyter --data-dir
 
 ### Windows
 
-**Important Note** - gomacro relies on the `plugin` package when importing third party libraries.  This package is only supported on Linux currently.  Thus, if you need to utilize third party packages in your Go notebooks and you are running on Windows, you should use the [Docker](#docker) install and run gophernotes/Jupyter in Docker.
+**Important Note** - gomacro relies on the `plugin` package when importing third party libraries.  This package is only supported on Linux and Mac OS X currently.  Thus, if you need to utilize third party packages in your Go notebooks and you are running on Windows, you should use the [Docker](#docker) install and run gophernotes/Jupyter in Docker.
 
 Make sure you have the MinGW toolchain:
 
@@ -188,7 +189,7 @@ $ docker run -it -p 8888:8888 -v /path/to/local/notebooks:/path/to/notebooks/in/
 
 gophernotes uses [gomacro](https://github.com/cosmos72/gomacro) under the hood to evaluate Go code interactively. You can evaluate most any Go code with gomacro, but there are some limitation, which are discussed in further detail [here](https://github.com/cosmos72/gomacro#current-status).  Most noteably, gophernotes does NOT support:
 
-- third party packages when running natively on Mac and Windows - This is a current limitation of the Go `plugin` package.
+- third party packages when running natively on Windows - This is a current limitation of the Go `plugin` package.
 - some corner cases on interpreted interfaces, as interface -&gt; interface type switch and type assertion, are not implemented yet.
 - conversion from typed constant to interpreted interface is not implemented. Workaround: assign the constant to a variable, then convert the variable to the interpreted interface type.
 - goto is only partially implemented.
