### 2018-06-29 397168e
doc: add travis badge
