### 2018-07-02 d76e9fb
update vendored gomacro to fix bug [gomacro#28](https://github.com/cosmos72/gomacro/issues/28): in 'import "path"' without an explicit name, use the name declared inside the package being imported, not paths.FileName(path)
