### 2018-07-09 d7999e6
Pin package repository and versions in Dockerfiles
