+++ Dockerfile
@@ -8,26 +8,30 @@ ADD . /go/src/github.com/gopherdata/gophernotes/
 RUN set -x \
     # install python and dependencies
     && apk update \
-    && apk --no-cache add \
+    && apk --no-cache \
+        --repository http://dl-4.alpinelinux.org/alpine/v3.7/community \
+        --repository http://dl-4.alpinelinux.org/alpine/v3.7/main \
+        --arch=x86_64 add \
         ca-certificates \
         python3 \
         su-exec \
         gcc \
         git \
         py3-zmq \
-        pkgconfig \ 
+        pkgconfig \
         zeromq-dev \
         musl-dev \
-    && pip3 install --upgrade pip \
+    && pip3 install --upgrade pip==9.0.3 \
     && ln -s /usr/bin/python3.5 /usr/bin/python \
     ## install Go
     && apk --update-cache --allow-untrusted \
         --repository http://dl-4.alpinelinux.org/alpine/edge/community \
         --arch=x86_64 add \
         go \
-    ## jupyter notebook 
+    ## jupyter notebook
     && ln -s /usr/include/locale.h /usr/include/xlocale.h \
-    && pip3 install jupyter notebook \
+    ### fix pyzmq to v16.0.2 as that is what is distributed with py3-zmq
+    && pip3 install jupyter notebook pyzmq==16.0.2 \
     ## install gophernotes
     && GOPATH=/go go install github.com/gopherdata/gophernotes \
     && cp /go/bin/gophernotes /usr/local/bin/ \
