### 2018-07-09 82426f8
Dockerfile python lib is now version 3.6 not 3.5
