### 2018-08-10 d0abe89
all: revert LICENSE to MIT
Fixes #126.
