### 2018-08-12 #508 @arokem
Hey @dwhitena: could I ask about the status of this submission? Have you had a chance to address the reviewer's comments?
