### 2018-08-13 #508 @dwhitena
@arokem Based on some of the feedback, I was hoping to have some of the plugin issues worked out before publishing. That way it would be more generally relevant. Since then, I haven't been able to devote much time to it, but @sbinet @spencerpark and @cosmos72 have taken the reigns. Maybe they can push this to the finish line?
