### 2018-08-18 df31119
Pin down tornado and ipykernel version to compatible ones. Fixes #137. (#140)
