### 2018-08-30 #508 @arokem
What is the status of this paper?

It would be great if these new contributors can push this along, but we'd like to know whether you intend to withdraw your paper. Otherwise, you can also consider adding some of these contributors as authors, if you believe that their contributions are crucial to completing the work required to meet the requirements for publication.
