### 2018-09-10 #508 @arfon
> @arfon Considering the author has indicated that he might not take this paper further to acceptance, should we close this issue for the time being, and reopen when the author(s) is(are) able to carry this work further?

Yes, let's close this.
